[#4161] fix datastore info test, move to extension
